Decoupling HTTP and core modules
Provides a way to integrate other clients such as gRPC or Hot Rod. Changes are backward-compatible on YAML level (with the exception of 'http.privatePools') but programmatic API has several breaking changes.
With the new model the session hosts destination tables, cache, request pools etc. as resources instead of fields.